Añadido módulo Factura

+++ src/main/java/app/negocio/factura/TFacturaProducto.java
@@ -5,11 +5,11 @@ public class TFacturaProducto {
 	private long idFactura;
 	private long idProducto;
 	private int cantidad;
-	private float precio;
+	private double precio;
 	
 	public TFacturaProducto() {}
 	
-	public TFacturaProducto(long idFactura, long idProducto, int cantidad, float precio) {
+	public TFacturaProducto(long idFactura, long idProducto, int cantidad, double precio) {
 		this.setIdFactura(idFactura);
 		this.setIdProducto(idProducto);
 		this.setCantidad(cantidad);
@@ -40,11 +40,11 @@ public class TFacturaProducto {
 		this.cantidad = cantidad;
 	}
 
-	public float getPrecio() {
+	public double getPrecio() {
 		return precio;
 	}
 
-	public void setPrecio(float precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
